refactor: share one BoundedLruMap policy between server and TUI
server/bounded-lru-map.ts and tui/src/generation-record-detail-cache.ts
each hand-rolled the same delete/reinsert/oldest-key LRU algorithm, with
diverging capacity validation and eviction behavior. Move the generic
policy to shared/bounded-lru-map.ts and have both the story graph cache
and the Generation Record detail cache build on it. The detail cache
keeps its narrow get/set interface and factory as the UI-facing seam;
only the storage swaps to the shared class.
